- removing testflight - fixing 64 bit errors and warnings (fixes #1084)
git-svn-id: http://svn.rtfnt.com/svn/studifutter/dev/trunk@288 dd375700-57de-49ca-a908-404e403163fd

+++ app/Studifutter/FHVersionUpdate.m
@@ -29,12 +29,12 @@
     NSArray *versionComponents1 = [version1 componentsSeparatedByString:@"."];
     NSArray *versionComponents2 = [version2 componentsSeparatedByString:@"."];
     
-    int numberOfComponents = 0;
+    NSUInteger numberOfComponents = 0;
     if ([versionComponents1 count] == [versionComponents2 count]) {
         numberOfComponents = [versionComponents1 count];
     } else {
-        int count1 = [versionComponents1 count];
-        int count2 = [versionComponents2 count];
+        NSUInteger count1 = [versionComponents1 count];
+        NSUInteger count2 = [versionComponents2 count];
         
         numberOfComponents = (count1 < count2) ? count1 : count2;
     }
